requirements.txt
Flask==2.0.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
